KAFKA-10195: Move offset management code out of ConsumerCoordinator
